Moved remaining unit tests
