+++ kernel/UnitTests/Framework.h
@@ -17,6 +17,17 @@ namespace UnitTests
      * @param apMessage The message to emit
      */
     void EmitTestSkipResult(char const* apMessage);
+
+    /**
+     * Runs kernel unit tests. Assumes MiniUART is set up for output and static
+     * constructors have been run
+     */
+    void Run();
+
+    /**
+     * Runs kernel unit tests to be run after static destructors are run
+     */
+    void RunPostStaticDestructors();
 }
 
 #endif // KERNEL_UNITTESTS_FRAMEWORK_H
